moved four main pages to folders
